Refactor Eye class and EnergyManager Changed Eye class to use the new EnergyManager class to manage energy consumption Fixed types in EnergyManager CalculateEnergyConsumptionEyeProcessing method

+++ Scenes/Entities/Creatures/Systems/Energy/EnergyManager.cs
@@ -74,7 +74,7 @@ public class EnergyManager
         return energyConsumed;
     }
 
-    public float CalculateEnergyConsumptionEyeProcessing(int EyeComplexity, int ActivatorPerEntity, float distance, int FOVVertical, int FOVHorizontal)
+    public float CalculateEnergyConsumptionEyeProcessing(int EyeComplexity, int ActivatorPerEntity, float distance, float FOVVertical, float FOVHorizontal)
     {
         float energyConsumed = EyeComplexity * ActivatorPerEntity * distance * (FOVVertical / 180) * (FOVHorizontal / 180) * EnergyCostEyeProcessing;
         return energyConsumed;
